Fix aggregates.livemd
As `bulk_create/4` returns in reversed order, the pattern matching didn't work as expected, resulting in the first ticket being closed instead of the last.

As this might be an implementation detail/ potentially prone to change, I suggest to make the query for the last ticket more explicit.

Revert naming changes

Update aggregates.livemd